+++ udata_front/theme/gouvfr/assets/js/.storybook/preview.ts
@@ -27,6 +27,7 @@ setup((app) => {
       children: [
         { path: '', component: Me, name: "me" },
         { path: 'datasets', component: Datasets, name: "me-datasets" },
+        { path: 'reuses', component: Reuses, name: "me-reuses" },
         { path: 'community-resources', component: CommunityRessources, name: "me-community-resources" },
       ]
     },