+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 - Update beta admin members page [#544](https://github.com/datagouv/udata-front/pull/544)
 - Fix Beta admin profile page [#546](https://github.com/datagouv/udata-front/pull/546)
 - Add beta admin user datasets page [#549](https://github.com/datagouv/udata-front/pull/549)
+- Add beta admin user's reuses page [#550](https://github.com/datagouv/udata-front/pull/550)
 
 ## 5.2.2 (2024-09-23)